[Feature] KVCache: enhance profiling
- add pyroscope based cpu profiling
- add env vars to control both cpu profiling and nvtx
